services/processing/requirements.txt
# services/processing/requirements.txt
pika==1.3.1
Pillow==10.0.0